move new endpoint to zks namespace
