+++ core/lib/zksync_core/src/api_server/web3/backend_jsonrpsee/namespaces/zks.rs
@@ -1,15 +1,11 @@
 use std::collections::HashMap;
+use itertools::Itertools;
 
-use zksync_types::{
-    api::{
-        BlockDetails, BridgeAddresses, L1BatchDetails, L2ToL1LogProof, Proof, ProtocolVersion,
-        TransactionDetails,
-    },
-    fee::Fee,
-    fee_model::FeeParams,
-    transaction_request::CallRequest,
-    Address, L1BatchNumber, L2BlockNumber, H256, U256, U64,
-};
+use zksync_types::{api::{
+    BlockDetails, BridgeAddresses, L1BatchDetails, L2ToL1LogProof, Proof, ProtocolVersion,
+    TransactionDetails,
+}, fee::Fee, fee_model::FeeParams, transaction_request::CallRequest, Address, L1BatchNumber, L2BlockNumber, H256, U256, U64, Bytes};
+use zksync_types::api::{ApiStorageLogQuery, ApiVmEvent, OptimisticTransactionResult};
 use zksync_web3_decl::{
     jsonrpsee::core::{async_trait, RpcResult},
     namespaces::zks::ZksNamespaceServer,
@@ -172,4 +168,32 @@ impl ZksNamespaceServer for ZksNamespace {
         self.get_base_token_l1_address_impl()
             .map_err(|err| self.current_method().map_err(err))
     }
+
+
+    async fn send_raw_transaction_optimistic(
+        &self,
+        tx_bytes: Bytes,
+    ) -> RpcResult<OptimisticTransactionResult> {
+        self.send_raw_transaction_optimistic_impl(tx_bytes)
+            .await
+            .map(|result| OptimisticTransactionResult {
+                transaction_hash: result.0,
+                storage_logs: result
+                    .1
+                    .logs
+                    .storage_logs
+                    .iter()
+                    .map(|m| ApiStorageLogQuery::from(m))
+                    .collect_vec(),
+                events: result
+                    .1
+                    .logs
+                    .events
+                    .iter()
+                    .map(|m| ApiVmEvent::from(m))
+                    .collect_vec(),
+            })
+            .map_err(|err| self.current_method().map_err(err))
+    }
+
 }
