+++ core/lib/zksync_core/src/api_server/web3/namespaces/zks.rs
@@ -1,25 +1,14 @@
 use std::{collections::HashMap, convert::TryInto};
 
 use anyhow::Context as _;
+use multivm::interface::VmExecutionResultAndLogs;
 use zksync_dal::{Connection, Core, CoreDal, DalError};
 use zksync_mini_merkle_tree::MiniMerkleTree;
 use zksync_system_constants::DEFAULT_L2_TX_GAS_PER_PUBDATA_BYTE;
-use zksync_types::{
-    api::{
-        BlockDetails, BridgeAddresses, GetLogsFilter, L1BatchDetails, L2ToL1LogProof, Proof,
-        ProtocolVersion, StorageProof, TransactionDetails,
-    },
-    fee::Fee,
-    fee_model::FeeParams,
-    l1::L1Tx,
-    l2::L2Tx,
-    l2_to_l1_log::{l2_to_l1_logs_tree_size, L2ToL1Log},
-    tokens::ETHEREUM_ADDRESS,
-    transaction_request::CallRequest,
-    utils::storage_key_for_standard_token_balance,
-    AccountTreeId, L1BatchNumber, L2BlockNumber, ProtocolVersionId, StorageKey, Transaction,
-    L1_MESSENGER_ADDRESS, L2_BASE_TOKEN_ADDRESS, REQUIRED_L1_TO_L2_GAS_PER_PUBDATA_BYTE, U256, U64,
-};
+use zksync_types::{api::{
+    BlockDetails, BridgeAddresses, GetLogsFilter, L1BatchDetails, L2ToL1LogProof, Proof,
+    ProtocolVersion, StorageProof, TransactionDetails,
+}, fee::Fee, fee_model::FeeParams, l1::L1Tx, l2::L2Tx, l2_to_l1_log::{l2_to_l1_logs_tree_size, L2ToL1Log}, tokens::ETHEREUM_ADDRESS, transaction_request::CallRequest, utils::storage_key_for_standard_token_balance, AccountTreeId, L1BatchNumber, L2BlockNumber, ProtocolVersionId, StorageKey, Transaction, L1_MESSENGER_ADDRESS, L2_BASE_TOKEN_ADDRESS, REQUIRED_L1_TO_L2_GAS_PER_PUBDATA_BYTE, U256, U64, Bytes};
 use zksync_utils::{address_to_h256, h256_to_u256};
 use zksync_web3_decl::{
     error::Web3Error,
@@ -30,6 +19,7 @@ use crate::api_server::{
     tree::TreeApiError,
     web3::{backend_jsonrpsee::MethodTracer, RpcState},
 };
+use crate::api_server::web3::metrics::API_METRICS;
 
 #[derive(Debug)]
 pub(crate) struct ZksNamespace {
@@ -570,5 +560,21 @@ impl ZksNamespace {
             .api_config
             .base_token_address
             .ok_or(Web3Error::NotImplemented)
+    }
+
+    #[tracing::instrument(skip(self, tx_bytes))]
+    pub async fn send_raw_transaction_optimistic_impl(
+        &self,
+        tx_bytes: Bytes,
+    ) -> Result<(H256, VmExecutionResultAndLogs), Web3Error> {
+        let (mut tx, hash) = self.state.parse_transaction_bytes(&tx_bytes.0)?;
+        tx.set_input(tx_bytes.0, hash);
+
+        let submit_result = self.state.tx_sender.submit_tx(tx).await;
+        submit_result.map(|result| (hash, result.1)).map_err(|err| {
+            tracing::debug!("Send raw transaction error: {err}");
+            API_METRICS.submit_tx_error[&err.prom_error_code()].inc();
+            err.into()
+        })
     }
 }
